Complete v0.0.1: harden foundation, resolve architectural risks, cut release
Chief Architect + Release Manager pass completing the v0.0.1 milestone.

Hardening (the two highest-severity findings from the architecture
review, both landed with tests proving they work, not just compile):
- Plugin ABI now carries an explicit version field and a forward-
  extension hook (ADR 0009 implemented, not just accepted); a real C
  fixture declaring a wrong version proves NativePluginLoader::load
  actually rejects it
- ECS component storage now requires Send + Sync, guarded by a
  compile-time assert_send_sync::<World>() test
- Entity generation widened u32 -> u64, moving wraparound from
  plausible-over-years to unreachable
- xtask check now detects and runs clippy instead of unconditionally
  skipping it, closing the local/CI parity gap

Two new architectural commitments, documented at the same depth as
rendering/physics/networking and equally unimplemented in this release:
- ADR 0011 + docs/architecture/ui-toolkit.md: CanaryUI, a trait-based
  UI abstraction bootstrapped on egui, shared between the editor and
  game-facing UI rather than built twice
- ADR 0012 + docs/architecture/state-and-versioning.md (Proposed):
  project state as an explicit, identifiable, versionable graph -- a
  founding principle so Git-friendly collaboration, marketplace
  packages, and eventual live editing become consequences of the
  architecture rather than separate features. Distinguishes runtime
  ECS entity identity from persistent/authored identity explicitly.
  Evaluated Automerge and Loro (pure-Rust CRDTs) as candidate backends
  for the long-term, genuinely-undecided collaboration mechanism.

Full repository audit:
- Every ADR (12) reviewed against reality; 0009 updated from
  decided-not-implemented to implemented
- Caught and fixed a pre-existing doc/code mismatch in
  plugin-system.md (documented a "register" hook that was never in
  the Plugin trait)
- 'Known limitations' sections in core-runtime.md and plugin-system.md
  updated: fixed items moved to 'resolved', not left as stale warnings
- Revised v0.0.1's own definition of done: archetype ECS, the Tier A
  WASM loader, real windowing, and change detection formally moved to
  v0.0.2+, with the reasoning for narrowing scope recorded rather than
  silently applied -- this project's versioning philosophy treats
  v0.0.1 as the first foundational milestone, not the first
  feature-complete one, and a bar that keeps growing every time a
  harder subsystem gets designed is a bar that never ships
- Every present-tense 'v0.0.1-pre1' doc reference updated to v0.0.1;
  historical references describing what specific past sessions
  produced deliberately left as-is

Release preparation:
- Workspace version bumped 0.0.1-pre1 -> 0.0.1
- CHANGELOG.md consolidated into one dated v0.0.1 entry
- RELEASE_NOTES_v0.0.1.md written: what this release is, deliberately
  is not, why its scope was set here, and v0.0.2 goals (archetype ECS
  migration, Tier A WASM loader, resolving ADR 0010 for real)
- docs/reviews/RELEASE_CHECKLIST.md: every checklist item verified,
  including the one this sandbox genuinely could not verify itself
  (clippy/CI, since neither rustup nor a GitHub Actions runner are
  reachable here) -- flagged honestly rather than assumed clean

16 tests passing across the workspace. cargo build, fmt --check, and
xtask check all clean.

+++ engine/canary-ecs/src/entity.rs
@@ -4,10 +4,25 @@
 /// match, so a stale handle to a despawned (and possibly index-recycled)
 /// entity is detectable rather than silently aliasing a new one. See
 /// `docs/architecture/core-runtime.md#ecs-architecture`.
+///
+/// `generation` is a `u64`, not a `u32`, on purpose: a `u32` generation
+/// wrapping on a single, very frequently recycled slot (a projectile
+/// spawn point over years of a long-lived server's uptime, say) was a
+/// real, if low-probability, correctness risk this project's own
+/// architecture review flagged (`docs/reviews/2026-08-senior-architecture-review.md`,
+/// Finding 4.2) — a wrapped generation could, in principle, collide with
+/// an old handle's generation for the same slot, silently aliasing two
+/// different entities. `u64` doesn't eliminate that possibility in
+/// theory, but it moves it from "plausible over a long-lived server's
+/// real uptime" to "not reachable by any realistic amount of runtime,"
+/// which is the standard this project holds correctness properties to.
+/// `index` stays `u32`: it's bounded by concurrent entity count, which
+/// doesn't have the same "recycled billions of times" exposure that made
+/// `generation` worth widening.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct Entity {
     pub(crate) index: u32,
-    pub(crate) generation: u32,
+    pub(crate) generation: u64,
 }
 
 impl Entity {
@@ -21,7 +36,7 @@ impl Entity {
 
     /// The generation of the slot this entity occupied at spawn time. See
     /// the type-level docs above.
-    pub fn generation(&self) -> u32 {
+    pub fn generation(&self) -> u64 {
         self.generation
     }
 }
